Added Search and Pagination functionality to get all businesses route
